Create test grid
First draft of a layout for a portfolio homepage. The grid is a row of
flex squares with minimal styling. On-hover, the square creates text
that superimposes over the squares and will show the project's title.

I'm realizing this repository should be rebranded into a series of case
studies for layouts.